ver porque no borrar, actualiza y añade

+++ tema8/ej11/php/deleteTarea.php
@@ -1,10 +1,16 @@
 <?php
-require_once("db.php");
+    require_once("db.php");
 
-$db = "nostas";
+    $db = "notas";
 
-$conexioin = conectar($db);
-$parametros = array(":id_nota" => $_POST["id_nota"]);
-$sql = "DELETE FROM notas WHERE id_nota = :id_nota";
-$pdo = $conexion->prepare($sql);
-$pdo->execute($parametros);
+    if(isset($_POST["eliminar"])){
+        $conexion = conectar($db);
+        $parametros = array(":id_nota" => $_POST["id_nota"]);
+        $sql = "DELETE FROM notas WHERE id_nota = :id_nota";
+        $pdo = $conexion->prepare($sql);
+        $pdo->execute($parametros);
+
+        $datos = $pdo->fetchAll(PDO::FETCH_ASSOC);
+        echo json_encode([message => "Borrado"]);
+    }
+?>
